+++ connector/queue/job.py
@@ -19,10 +19,11 @@
 #
 ##############################################################################
 
-import sys
-import logging
 import inspect
+import functools
+import logging
 import uuid
+import sys
 from datetime import datetime, timedelta, MINYEAR
 from pickle import loads, dumps, UnpicklingError
 
@@ -598,6 +599,11 @@ class Job(object):
         if result is not None:
             self.result = result
 
+    def related_action(self, session):
+        if not hasattr(self.func, 'related_action'):
+            return None
+        return self.func.related_action(session, self)
+
 
 def job(func):
     """ Decorator for jobs.
@@ -630,7 +636,7 @@ def job(func):
                     infinite retries. Default is 5.
      * eta: the job can be executed only after this datetime
             (or now + timedelta if a timedelta or integer is given)
-     
+
      * description : a human description of the job,
                      intended to discriminate job instances
                      (Default is the func.__doc__ or 'Function %s' % func.__name__)
@@ -655,6 +661,9 @@ def job(func):
         # => the job will be executed with a low priority and not before a
         # delay of 5 hours from now
 
+    See also: :py:func:`related_action` a related action can be attached
+    to a job
+
     """
     def delay(session, model_name, *args, **kwargs):
         """Enqueue the function. Return the uuid of the created job."""
@@ -665,3 +674,61 @@ def job(func):
             **kwargs)
     func.delay = delay
     return func
+
+
+def related_action(action=lambda session, job: None, **kwargs):
+    """ Attach a *Related Action* to a job.
+
+    A *Related Action* will appear as a button on the OpenERP view.
+    The button will execute the action, usually it will open the
+    form view of the record related to the job.
+
+    The ``action`` must be a callable that responds to arguments::
+
+        session, job, **kwargs
+
+    Example usage:
+
+    .. code-block:: python
+
+        def related_action_partner(session, job):
+            model = job.args[0]
+            partner_id = job.args[1]
+            # eventually get the real ID if partner_id is a binding ID
+            action = {
+                'name': _("Partner"),
+                'type': 'ir.actions.act_window',
+                'res_model': model,
+                'view_type': 'form',
+                'view_mode': 'form',
+                'res_id': partner_id,
+            }
+            return action
+
+        @job
+        @related_action(action=related_action_partner)
+        def export_partner(session, model_name, partner_id):
+            # ...
+
+    The kwargs are transmitted to the action:
+
+    .. code-block:: python
+
+        def related_action_product(session, job, extra_arg=1):
+            assert extra_arg == 2
+            model = job.args[0]
+            product_id = job.args[1]
+
+        @job
+        @related_action(action=related_action_product, extra_arg=2)
+        def export_product(session, model_name, product_id):
+            # ...
+
+    """
+    def decorate(func):
+        if kwargs:
+            func.related_action = functools.partial(action, **kwargs)
+        else:
+            func.related_action = action
+        return func
+    return decorate